oat/pyproject.toml
[project]
name = "oat"
version = "0.1.0"
description = "OAT: Ordered Action Tokenization"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
dependencies = [
    "einops>=0.8.1",
    "hydra-core>=1.3.2",
    "wandb>=0.24.0",
    "dill>=0.4.1",
    "zarr>=2.18.3",
    "numba>=0.63.1",
    "diffusers>=0.36.0",
    "accelerate>=1.12.0",
    "transformers>=4.57.6",
    "av>=16.1.0",
    "gymnasium>=1.2.3",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "vector-quantize-pytorch>=1.27.19",
    "robomimic==0.2.0",
    "robosuite==1.4.0",
    "easydict>=1.13",
    "bddl>=3.6.0",
    "cloudpickle>=3.1.2",
    "libero",
    "matplotlib>=3.10.8",
    "gym>=0.26.2",
    "pandas>=2.3.3",
]

[project.urls]
Repository = "https://github.com/Chaoqi-LIU/oat"

[tool.uv.sources]
libero = { workspace = true, editable = true }

[tool.uv.workspace]
members = [
    "third_party/LIBERO",
]

[tool.hatch.build.targets.wheel]
packages = ["oat"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
